docs(tvix): mention channels separately
Returns the original (pre community split) channel into a more prominent
position, and explicitly mentions that #tvix-dev is for the community
centered around the new store model.

This relates to b/431, and is part of the transition setup while we
figure out how to organise the project split.

Change-Id: I0d8c5e86c914f51db222b054d51038e2023281d6
Reviewed-on: https://cl.tvl.fyi/c/depot/+/12950
Tested-by: BuildkiteCI
Reviewed-by: Profpatsch <mail@profpatsch.de>
Autosubmit: tazjin <tazjin@tvl.su>