docs(REVIEWS): link additional related documentation
Unfortunately, the git book doesn't link the man pages of the involved
tools, but I assume users will figure out to read them on their own.
I've not linked https://git-send-email.io due to the slight reservation
that it's not an official resource (and is an advertisment), but I don't
feel strongly on that. The linked chapter of the Git Book links it in a
footnote anyways.

Change-Id: I70e3c75396006256c775f84bc7aeb5a4bde081a1
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13032
Tested-by: BuildkiteCI
Reviewed-by: tazjin <tazjin@tvl.su>
Autosubmit: sterni <sternenseemann@systemli.org>