feat(sterni/blipqn): actually rethrow error messages in WithFlipdot
Passing a namespace as 𝕩 to •CurrentError is not allowed since it is
reserved for future use (not sure how it'll change the behavior
exactly). Since WithFlipdot always passes a namespace as 𝕩, this is a
relevant problem for blipqn. To work around it, we wrap the passed
argument in a list (and unwrap it for 𝔾).

Change-Id: I3382f754cbb64ad799e47b7d174d641c43a0f2e0
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13125
Reviewed-by: sterni <sternenseemann@systemli.org>
Autosubmit: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI