fix(tvix/cli): NIX_PATH contents are duplicated when -I is not passed
Fix a bug introduced in 2daa4832490eddcfb2b9a19d7f651252ec766062

When the -I or --extra-nix-path option is not passed, the cli will
interpret the contents of the NIX_PATH environment variable twice:
once as part of the env attribute of clap's derive macro, and again in a
custom nix_path function. The custom function implements the appending
behavior nix users likely expect (see commit mentioned above), so the
env attribute is dropped. The downside is that NIX_PATH is no longer
rendered in the cli's help text. We can reclaim this behavior if we want
to, but that's left for later.

Change-Id: I2c0939d9453ea1af73e9ec15e45b25bf035a7493
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13207
Autosubmit: benjaminedwardwebb <benjaminedwardwebb@gmail.com>
Tested-by: BuildkiteCI
Reviewed-by: flokli <flokli@flokli.de>