feat(users/Profpatsch/whatcd-resolver): simplify search
Throws out htmx for search in favor of a full page-reload for search.
This means we can reload the page and it will re-do the search.

It’s not POST anymore cause we don’t really change anything, just
do some caching when doing the search. Could probably also cache the
search result, but simple simple

Change-Id: I70b0cbac55f67c986aeecfedcf0b3968890b3b9c
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13212
Tested-by: BuildkiteCI
Reviewed-by: Profpatsch <mail@profpatsch.de>