feat(snix/castore): builtins.fetchGit
This adds builtins.fetchgit and the backing git import based on
gitoxide.

The implementation uses a temporary directory to clone a bare git
repository in a separate thread. gitoxide does not yet have full async
support for cloning ssh-based repositories, which is why a separate
thread is used.

builtins.fetchgit has more requirements than the Nix equivalent. rev is
always required and a sha256 to check is not supported at this time.

Change-Id: I035756c300bd694ab4ca4640ba941e4603925523