fix(snix/build): use try_into_name_and_node()
The BuildRequest trait currently still uses the proto Build struct
as a response type, and that one uses nodes with names, causing
.try_into_anonymous_node() to panic.

Use try_into_name_and_node(), and discard the name for now. We should
update the trait to use a stricter type for this.

Change-Id: Ib183a163e885443c58c42808e464d0611eaae324
Reviewed-on: https://cl.snix.dev/c/snix/+/30169
Reviewed-by: Vova Kryachko <v.kryachko@gmail.com>
Tested-by: besadii
Autosubmit: Florian Klink <flokli@flokli.de>