feat(ops/modules/monorepo-gerrit): re-enable Code Owners
Code Owners was disabled because it's very sensitive about the identity
of the committers and while pushing the original history, this was a
distraction.

Now that the history has been pushed and everyone is back to their
normal identity, it's fine to enable it again.

Fixes #83.

Change-Id: I4181d6af4eca489d4827b1c1ee606dfbb28a05c9
Reviewed-on: https://cl.snix.dev/c/snix/+/30173
Tested-by: besadii
Reviewed-by: Florian Klink <flokli@flokli.de>
Autosubmit: Ryan Lahfa <masterancpp@gmail.com>