fix(nix-compat-derive-tests): Get rid of compile-tests flag
Because we are now using mkCargoBuild to build this crate in CI we can
get rid of the compile-tests flag and always run the compile tests.
This is because the trybuild crate works when using the the same build
method we used for rustdoc an clippy.

Change-Id: I3f168a987a09bf6f3e3e7c1ab3aaa1acd9d010cf
Reviewed-on: https://cl.snix.dev/c/snix/+/30651
Reviewed-by: Florian Klink <flokli@flokli.de>
Autosubmit: Brian Olsen <brian@maven-group.org>
Tested-by: besadii